Scope down import: URLs
This closes #75 by acknowledging that we won't be able to solve that problem.

import: URLs become less primary to the formulation of the spec; it is no longer about controlling import: URL resolution, with module specifier resolution being a natural extension of that, but instead is about changing module specifier resolution, with import: URLs as an additional feature that builds on module specifier resolution.

import: URLs no longer involve URL parser changes, as #75 explains how making URL parsing sensitive to the active script conflicts badly with URLs' nature as string-serializable things which can be passed around. Instead, we retreat to a position of import: URL fetches being relative to the request's referrer, i.e. using the request's referrer to resolve ./ specifiers or determine which scopes apply. This will hopefully work better.

We also note in the proto-spec explicitly our plans to first spec without fallback between fetch schemes.