Discard addresses with missing trailing slashes
If an entry has an address with no trailing slash, but the specifier key has a trailing slash, then discard the address. Closes #93.

This also introduces warning infrastructure, as discussed in #80, and includes our first warning, for when such addresses are discarded.

A related piece of trailing-slash behavior is being discussed in #91; stay tuned for further tweaks.