Restructure resolveImportsMatch algorithm
This change has no observable effects to our tests, but makes it clear that there is a logic hole in the package prefix-match case with empty address arrays (which was not covered by tests, apparently). The following commit will fix that.

This also restructures the code in a way that is likely to work better for adding built-in module resolution.