Make import maps registration priorities and error events in-order
Previously, import maps were processed when they are fetched, and therefore the priority of import maps and order of error events was nondeterministic.

This commit introduces a mechanism to process import maps in the order of "prepare a script", and thus fixes #114. To do so, it expands the general integration of import maps with "prepare a script" to make them more symmetrical to module and classic scripts in that regard.