Slim down to a minimal core
This re-focuses the import maps explainer and specification on a minimal core which is more likely to be implementable and agreeable in the near future. It removes multiple import map support, fallback support, and built-in module support for now. The explainer contains a "Further work" section which details these and other potential future work.

In terms of normative changes:

* The parser now rejects non-string addresses, so arrays and null are no longer valid
* The resolver becomes much simpler due to this change, as well as the removal of built-in module support
* There is no more procedure for merging import maps
* Script fetches only wait on a single "pending import map script" instead of a "list of pending import map scripts"