Remove advisements about future Fetch-based version
I think we should set aside that idea, and close #169. It has serious problems with establishing the appropriate base URL, as discussed in previous issues. So we can remove these paragraphs from the spec and avoid distracting readers with the idea that the spec might dramatically change in the future.

This also removes the indication that we modify "fetch a module worker script graph". It's not clear how we'll integrate in the future, but it seems unlikely it would be via this exact mechanism.